refactor(utils): move util classes into its own component directory
