+++ components/schedule/class/ClassPopularityMeter.tsx
@@ -4,7 +4,7 @@ import { ClassPopularity, StatusColors } from "../../../types/rezervoTypes";
 import { SitClass } from "../../../types/sitTypes";
 import { isClassInThePast } from "../../../lib/iBooking";
 import { stringifyClassPopularity, determineClassPopularity } from "../../../lib/popularity";
-import RippleBadge from "../../RippleBadge";
+import RippleBadge from "../../utils/RippleBadge";
 import { Tooltip } from "@mui/material";
 
 const ClassPopularityIcon = ({ popularity }: { popularity: ClassPopularity }) => {
