+++ pages/index.tsx
@@ -10,7 +10,7 @@ import MobileConfigUpdateBar from "../components/MobileConfigUpdateBar";
 import { createClassPopularityIndex } from "../lib/popularity";
 import { DateTime } from "luxon";
 import { useUserConfig } from "../hooks/useUserConfig";
-import PageHead from "../components/PageHead";
+import PageHead from "../components/utils/PageHead";
 import ClassInfoModal from "../components/modals/ClassInfo/ClassInfoModal";
 import AgendaModal from "../components/modals/Agenda/AgendaModal";
 import SettingsModal from "../components/modals/Settings/SettingsModal";
